parser/requirements.txt
anthropic>=0.40.0
pdf2image>=1.17.0
Pillow>=10.0.0
